move infer.go to types_go118.go
